pyproject.toml
[tool.poetry]
name = "rdsa-utils"
version = "0.1.0-dev0"
description = ""
authors = ["Reproducible Data Science & Analysis, ONS"]
maintainers = [
    "Matthew Price <Matthew.G.Price@ons.gov.uk>",
    "Meg Scammell <Meg.Scammell@ons.gov.uk>",
    "Dominic Bean <Dominic.Bean@ons.gov.uk>",
]

readme = "README.md"
repository = "https://github.com/ONSdigital/rdsa-utils"
documentation = "https://github.com/ONSdigital/rdsa-utils"


[tool.poetry.dependencies]
python = ">=3.8, <3.11"

cloudpathlib = {extras = ["gs"], version = "^0.15.1"}
humanfriendly = "^9.1"
more-itertools = "^9.0.0"
pandas = "1.5.3"
pydantic = "2.0"
pyspark = "3.3.1"
pyyaml = "^6.0.1"
tomli = "^2.0.1"

[tool.poetry.group.gcp-local.dependencies]
google-cloud-bigquery = "3.3.5"
google-cloud-storage = "2.5.0"

[tool.poetry.group.test.dependencies]
# All test dependencies.
chispa = "^0.9.2"
coverage = {extras = ["toml"], version = "^7.1.0"}
pytest = "^7.1.0"
pytest-cov = "^4.0.0"
pytest-lazy-fixture = "^0.6.0"
pytest-mock = "^3.8.0"

[tool.poetry.group.dev.dependencies]
# These are purely used for development tools only.
bump2version = "^1.0.0"
pre-commit = "^2.6.0"
ruff = "^0.0.270"

[tool.poetry.group.doc.dependencies]
# Documentation
mkdocs = "^1.4.2"
mkdocs-tech-docs-template = "^0.0.25"
mkdocstrings = {extras = ["python"], version = "^0.22.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
select = [
   "A",    # https://beta.ruff.rs/docs/rules/#flake8-builtins-a
   "B",    # https://beta.ruff.rs/docs/rules/#flake8-bugbear-b
   "C4",   # https://beta.ruff.rs/docs/rules/#flake8-comprehensions-c4
   "COM",  # https://beta.ruff.rs/docs/rules/#flake8-commas-com
   "D",    # https://beta.ruff.rs/docs/rules/#pydocstyle-d
   "DTZ",  # https://beta.ruff.rs/docs/rules/#flake8-datetimez-dtz
   "E",    # https://beta.ruff.rs/docs/rules/#error-e
   "EM",   # https://beta.ruff.rs/docs/rules/#flake8-errmsg-em
   "F",    # https://beta.ruff.rs/docs/rules/#pyflakes-f
   "ISC",  # https://beta.ruff.rs/docs/rules/#flake8-implicit-str-concat-isc
   "N",    # https://beta.ruff.rs/docs/rules/#pep8-naming-n
   "NPY",  # https://beta.ruff.rs/docs/rules/#numpy-specific-rules-npy
   "PD",   # https://beta.ruff.rs/docs/rules/#pandas-vet-pd
   "PT",   # https://beta.ruff.rs/docs/rules/#flake8-pytest-style-pt
   "PTH",  # https://beta.ruff.rs/docs/rules/#flake8-use-pathlib-pth
   "Q",    # https://beta.ruff.rs/docs/rules/#flake8-quotes-q
   "T10",  # https://beta.ruff.rs/docs/rules/#flake8-debugger-t10
   "T20",  # https://beta.ruff.rs/docs/rules/#flake8-print-t20
   "W",    # https://beta.ruff.rs/docs/rules/#warning-w
   "YTT",  # https://beta.ruff.rs/docs/rules/#flake8-2020-ytt
]

# Ignore these rules everywhere.
ignore = [
   # N812: lowercase-imported-as-non-lowercase
   # This clashes with pyspark types and functions imports.
   "N812",
   # PD901: pandas-df-variable-name
   # This complains about anything called df.
   "PD901",
   # Clashes with Pyspark comparison with False.
   "E712",
]
line-length = 80

exclude = ["dev_utils/*"]

[tool.ruff.pydocstyle]
convention = "numpy"

[tool.ruff.flake8-quotes]
inline-quotes = "single"

[tool.ruff.per-file-ignores]
# D104 Missing docstring in public package.
# B006 Need to set {} as default argument.
# PTH123 Need to set to open().
# B008 relates to T.BooleanType(). This is required in the
# convert_cols_to_struct_col function in helpers_spark.py.
# E202, E203 is whitespace before/after (), which is sometimes
# broken when aligning data in test scripts. Not implemented in Ruff.
# E501 is line too long.
# F403 and F405 relate to importing using *. It is usually useful to
# do so in test scripts.
"*/__init__.py" = [
   "D104",
]
"*/" = [
   "B006",
   "PTH123",
   "B008",
]
"tests/*" = [
   "ANN",
   "D100",
   # "E202", "E203",
   "E501",
   "F403", "F405",
   "PT011",
   "B017",
]

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]